some tests passing
